refactor: move function to separate file to resolve circular dependency
