+++ src/utils.ts
@@ -8,16 +8,13 @@ import {
   Field,
   Poseidon,
   ProvableType,
-  Proof,
 } from 'o1js';
 import { Bigint2048 } from './rsa.js';
 import pako from 'pako';
-import { BLOCKS_PER_BASE_PROOF, hashProgram } from './recursion.js';
 import { DynamicArray, StaticArray } from 'mina-attestations';
 import {
   Block32,
   BlockBytes,
-  MerkleBlocks,
   State32,
   WordBytes,
 } from './dataTypes.js';
@@ -32,8 +29,6 @@ export {
   digitBytesToTimestamp,
   digitBytesToInt,
   commitBlock256,
-  hashBlocks,
-  hashBlock256,
   state32ToBytes
 };
 
@@ -483,78 +478,6 @@ function toInput(block: Block32) {
 
   // Otherwise, fall back to using toFields and wrap the result
   return { fields: type.toFields(block) };
-}
-
-/**
- * Recursively hashes a sequence of Merkle blocks using a proof system.
- *
- * This function splits the input blocks into two parts:
- * - A "tail" of blocks to be hashed directly.
- * - A "remaining" prefix to be hashed recursively or using a base method.
- *
- * The result is a SHA-256 style hash after applying all block transformations.
- *
- * @param {MerkleBlocks} blocks - The full array of Merkle blocks to hash.
- * @param {{ blocksInThisProof: number }} options - The number of blocks to include in the current proof.
- * @returns {Promise<State32>} The resulting state after hashing all blocks.
- * @notice - Taken from https://github.com/zksecurity/mina-attestations/blob/835d8d47566c4c065fa34c88af7ce99a5993425c/src/email/zkemail.ts#L210
- */
-async function hashBlocks(
-  blocks: MerkleBlocks,
-  options: { blocksInThisProof: number }
-): Promise<State32> {
-  let { blocksInThisProof } = options;
-
-  // split blocks into remaining part and final part
-  // the final part is done in this proof, the remaining part is done recursively
-  let { remaining, tail } = MerkleBlocks.popTail(blocks, blocksInThisProof);
-
-  // recursively hash the first, "remaining" part
-  let proof = await Provable.witnessAsync(hashProgram.Proof, async () => {
-    // optionally disable the inner proof
-
-    // convert the blocks to constants
-    let blocksForProof = Provable.toConstant(MerkleBlocks, remaining.clone());
-
-    // figure out if we can call the base method or need to recurse
-    let nBlocksRemaining = remaining.lengthUnconstrained().get();
-    let proof: Proof<MerkleBlocks, State32>;
-
-    if (nBlocksRemaining <= BLOCKS_PER_BASE_PROOF) {
-      console.log({ nBlocksRemaining, method: 'hashBase' });
-      ({ proof } = await hashProgram.hashBase(blocksForProof));
-    } else {
-      console.log({ nBlocksRemaining, method: 'hashRecursive' });
-      ({ proof } = await hashProgram.hashRecursive(blocksForProof));
-    }
-    return proof;
-  });
-  proof.declare();
-  proof.verify();
-
-  // constrain public input to match the remaining blocks
-  remaining.hash.assertEquals(proof.publicInput.hash);
-
-  // continue hashing the final part
-  let state = proof.publicOutput;
-  tail.forEach(({ isSome, value: block }) => {
-    let nextState = hashBlock256(state, block);
-    state = Provable.if(isSome, State32, nextState, state);
-  });
-  return state;
-}
-
-/**
- * Computes the SHA-256 hash of a block using a given initial state.
- *
- * @param {State32} state - The initial SHA-256 state (8 UInt32s).
- * @param {Block32} block - The message block to hash (16 UInt32s).
- * @returns {State32} The new SHA-256 state after compression.
- * @notice - Taken from https://github.com/zksecurity/mina-attestations/blob/main/src/dynamic/dynamic-sha2.ts#L511
- */
-function hashBlock256(state: State32, block: Block32): State32 {
-  let W = Gadgets.SHA2.messageSchedule(256, block.array);
-  return State32.from(Gadgets.SHA2.compression(256, state.array, W));
 }
 
 /**
